PLT-0019 , PLT-0020

+++ PilotSmithApp.UserInterface/Controllers/EnquiryController.cs
@@ -328,7 +328,7 @@ namespace PilotSmithApp.UserInterface.Controllers
         public ActionResult CheckQty(decimal Qty)
         {
             
-            if ((int)Qty == 0)
+            if (Qty == 0)
             {
 
                 return Json("<p><span style='vertical-align: 2px'>Value could not be zero!</span></p>", JsonRequestBehavior.AllowGet);
@@ -343,7 +343,7 @@ namespace PilotSmithApp.UserInterface.Controllers
         public ActionResult CheckRate( decimal Rate)
         {
             
-            if ((int)Rate == 0)
+            if (Rate == 0)
             {
 
                 return Json("<p><span style='vertical-align: 2px'>Value could not be zero!</span></p>", JsonRequestBehavior.AllowGet);
