+++ PilotSmithApp.UserInterface/Controllers/QuotationController.cs
@@ -535,7 +535,7 @@ namespace PilotSmithApp.UserInterface.Controllers
         public ActionResult CheckQty(decimal Qty)
         {
 
-            if ((int)Qty == 0)
+            if (Qty == 0)
             {
 
                 return Json("<p><span style='vertical-align: 2px'>Value could not be zero!</span></p>", JsonRequestBehavior.AllowGet);
@@ -550,7 +550,7 @@ namespace PilotSmithApp.UserInterface.Controllers
         public ActionResult CheckRate(decimal Rate)
         {
 
-            if ((int)Rate == 0)
+            if (Rate == 0)
             {
 
                 return Json("<p><span style='vertical-align: 2px'>Value could not be zero!</span></p>", JsonRequestBehavior.AllowGet);
